+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Admin field that for admin users
 - ability for admins to delete posts
 - ability to upload file at `/upload` with the name profilePicture
+- publicPath to config file to configure the directory for public files
 
 ### Removed